+++ app/src/main/java/org/stepic/droid/di/course_general/CourseGeneralComponent.kt
@@ -6,7 +6,9 @@ import org.stepic.droid.di.course_list.CourseGeneralScope
 import org.stepic.droid.di.course_list.CourseListComponent
 
 @CourseGeneralScope
-@Subcomponent(modules = arrayOf(CourseGeneralModule::class))
+@Subcomponent(modules = [
+    CourseGeneralModule::class
+])
 interface CourseGeneralComponent {
 
     @Subcomponent.Builder
